bioinformatics_dna_count/requirements.txt
altair==4.1.0
pandas==1.3.5
Pillow==9.0.0
streamlit==1.3.1
